// src/lib/http/generated/models/list-animal-from-user-response-dto-output-items-item.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API de Cuidados com Animais
 * OpenAPI spec version: 1.0
 */
import type { ListAnimalFromUserResponseDtoOutputItemsItemAge } from "./list-animal-from-user-response-dto-output-items-item-age";
import type { ListAnimalFromUserResponseDtoOutputItemsItemAsset } from "./list-animal-from-user-response-dto-output-items-item-asset";
import type { ListAnimalFromUserResponseDtoOutputItemsItemAssetId } from "./list-animal-from-user-response-dto-output-items-item-asset-id";
import type { ListAnimalFromUserResponseDtoOutputItemsItemBreed } from "./list-animal-from-user-response-dto-output-items-item-breed";
import type { ListAnimalFromUserResponseDtoOutputItemsItemBreedId } from "./list-animal-from-user-response-dto-output-items-item-breed-id";
import type { ListAnimalFromUserResponseDtoOutputItemsItemWeight } from "./list-animal-from-user-response-dto-output-items-item-weight";

export type ListAnimalFromUserResponseDtoOutputItemsItem = {
	id: string;
	name: string;
	age?: ListAnimalFromUserResponseDtoOutputItemsItemAge;
	weight?: ListAnimalFromUserResponseDtoOutputItemsItemWeight;
	userId: string;
	assetId?: ListAnimalFromUserResponseDtoOutputItemsItemAssetId;
	breedId?: ListAnimalFromUserResponseDtoOutputItemsItemBreedId;
	breed: ListAnimalFromUserResponseDtoOutputItemsItemBreed;
	asset?: ListAnimalFromUserResponseDtoOutputItemsItemAsset;
};